requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0